# docs/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Autograder API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/staff/admin/courses": {
      "get": {
        "tags": [
          "staff-admin"
        ],
        "summary": "Get Courses Admin",
        "operationId": "get_courses_admin_staff_admin_courses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CourseAdminDetail"
                  },
                  "type": "array",
                  "title": "Response Get Courses Admin Staff Admin Courses Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "staff-admin"
        ],
        "summary": "Create Course Admin",
        "operationId": "create_course_admin_staff_admin_courses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseAdminDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/staff/admin/courses/{course_id}": {
      "put": {
        "tags": [
          "staff-admin"
        ],
        "summary": "Update Course Admin",
        "operationId": "update_course_admin_staff_admin_courses__course_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseAdminDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "staff-admin"
        ],
        "summary": "Deactivate Course Admin",
        "operationId": "deactivate_course_admin_staff_admin_courses__course_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/staff/admin/courses/{course_id}/sections": {
      "get": {
        "tags": [
          "staff-admin"
        ],
        "summary": "Get Sections Admin",
        "operationId": "get_sections_admin_staff_admin_courses__course_id__sections_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectionAdminDetail"
                  },
                  "title": "Response Get Sections Admin Staff Admin Courses  Course Id  Sections Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "staff-admin"
        ],
        "summary": "Create Section Admin",
        "operationId": "create_section_admin_staff_admin_courses__course_id__sections_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SectionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectionAdminDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sandbox/courses/{course_id}/assignments": {
      "get": {
        "tags": [
          "sandbox"
        ],
        "summary": "List Assignments",
        "operationId": "list_assignments_sandbox_courses__course_id__assignments_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "x-sandbox-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sandbox-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxAssignmentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/staff/courses": {
      "get": {
        "tags": [
          "staff-courses"
        ],
        "summary": "Get Staff Courses",
        "operationId": "get_staff_courses_staff_courses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffCourseListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/staff/courses/{course_id}/assignments": {
      "get": {
        "tags": [
          "staff-courses"
        ],
        "summary": "Get Staff Assignments",
        "operationId": "get_staff_assignments_staff_courses__course_id__assignments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffAssignmentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "staff-assignments"
        ],
        "summary": "Create New Assignment",
        "operationId": "create_new_assignment_staff_courses__course_id__assignments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffAssignmentSetup"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/staff/courses/{course_id}/sections": {
      "get": {
        "tags": [
          "staff-courses"
        ],
        "summary": "Get Staff Sections",
        "operationId": "get_staff_sections_staff_courses__course_id__sections_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffSectionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/staff/courses/{course_id}/concepts": {
      "get": {
        "tags": [
          "staff-courses"
        ],
        "summary": "Get Concepts",
        "operationId": "get_concepts_staff_courses__course_id__concepts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseConceptsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "staff-courses"
        ],
        "summary": "Update Concepts",
        "operationId": "update_concepts_staff_courses__course_id__concepts_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseConceptsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseConceptsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignmentCreate": {
        "properties": {
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9_-]*$",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "python"
          },
          "canvas_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canvas Ref"
          },
          "sandbox_enabled": {
            "type": "boolean",
            "title": "Sandbox Enabled",
            "default": true
          },
          "module_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Id"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "title"
        ],
        "title": "AssignmentCreate"
      },
...
      "CourseAdminDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "term": {
            "type": "string",
            "title": "Term"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "instructor_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructor Id"
          },
          "instructor_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructor Email"
          },
          "ia_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ia Id"
          },
          "ia_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ia Email"
          },
          "default_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Default Concepts"
          },
          "section_count": {
            "type": "integer",
            "title": "Section Count"
          },
          "assignment_count": {
            "type": "integer",
            "title": "Assignment Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "title",
          "term",
          "is_active",
          "default_concepts",
          "section_count",
          "assignment_count"
        ],
        "title": "CourseAdminDetail"
      },
      "CourseConceptsResponse": {
        "properties": {
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "default_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Default Concepts"
          },
          "modules": {
            "items": {
              "$ref": "#/components/schemas/ModuleConfig"
            },
            "type": "array",
            "title": "Modules"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "default_concepts"
        ],
        "title": "CourseConceptsResponse"
      },
      "CourseConceptsUpdate": {
        "properties": {
          "default_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Default Concepts"
          },
          "modules": {
            "items": {
              "$ref": "#/components/schemas/ModuleConfig"
            },
            "type": "array",
            "title": "Modules"
          }
        },
        "type": "object",
        "required": [
          "default_concepts"
        ],
        "title": "CourseConceptsUpdate"
      },
      "CourseCreate": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "term": {
            "type": "string",
            "title": "Term"
          },
          "default_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Default Concepts",
            "default": []
          },
          "instructor_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructor Email"
          },
          "instructor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructor Name"
          },
          "ia_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ia Email"
          },
          "ia_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ia Name"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "term"
        ],
        "title": "CourseCreate"
      },
      "CourseUpdate": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "term": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term"
          },
          "default_concepts": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Concepts"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "instructor_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructor Id"
          },
          "ia_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ia Id"
          },
          "instructor_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructor Email"
          },
          "instructor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructor Name"
          },
          "ia_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ia Email"
          },
          "ia_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ia Name"
          }
        },
        "type": "object",
        "title": "CourseUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SandboxAssignmentListResponse": {
        "properties": {
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/SandboxAssignmentSummary"
            },
            "type": "array",
            "title": "Assignments"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "assignments"
        ],
        "title": "SandboxAssignmentListResponse"
      },
...
      "SectionAdminDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "course_id": {
            "type": "integer",
            "title": "Course Id"
          },
          "crn": {
            "type": "string",
            "title": "Crn"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "crn",
          "is_active"
        ],
        "title": "SectionAdminDetail"
      },
      "SectionCreate": {
        "properties": {
          "crn": {
            "type": "string",
            "title": "Crn"
          }
        },
        "type": "object",
        "required": [
          "crn"
        ],
        "title": "SectionCreate"
      },
...
      "StaffAssignmentListResponse": {
        "properties": {
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/StaffAssignmentSummary"
            },
            "type": "array",
            "title": "Assignments"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "assignments"
        ],
        "title": "StaffAssignmentListResponse"
      },
      "StaffAssignmentSetup": {
        "properties": {
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "assignment_id": {
            "type": "string",
            "title": "Assignment Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "sandbox_enabled": {
            "type": "boolean",
            "title": "Sandbox Enabled"
          },
          "base_points": {
            "type": "integer",
            "title": "Base Points"
          },
          "extra_credit_points": {
            "type": "integer",
            "title": "Extra Credit Points"
          },
          "entrypoint_path": {
            "type": "string",
            "title": "Entrypoint Path"
          },
          "scoring_items": {
            "items": {
              "$ref": "#/components/schemas/ScoringItem"
            },
            "type": "array",
            "title": "Scoring Items"
          },
          "rubric_groups": {
            "items": {
              "$ref": "#/components/schemas/RubricGroup"
            },
            "type": "array",
            "title": "Rubric Groups"
          },
          "completion_requirements": {
            "items": {
              "$ref": "#/components/schemas/CompletionRequirement"
            },
            "type": "array",
            "title": "Completion Requirements"
          },
          "canvas_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canvas Ref"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactMetadata"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "config_json": {
            "$ref": "#/components/schemas/AssignmentConfigV1"
          },
          "module_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Id"
          },
          "effective_allowed_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Effective Allowed Concepts"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "assignment_id",
          "title",
          "language",
          "sandbox_enabled",
          "base_points",
          "extra_credit_points",
          "entrypoint_path",
          "scoring_items",
          "rubric_groups",
          "completion_requirements",
          "artifacts",
          "config_json"
        ],
        "title": "StaffAssignmentSetup"
      },
...
      "StaffCourseListResponse": {
        "properties": {
          "courses": {
            "items": {
              "$ref": "#/components/schemas/StaffCourseSummary"
            },
            "type": "array",
            "title": "Courses"
          }
        },
        "type": "object",
        "required": [
          "courses"
        ],
        "title": "StaffCourseListResponse"
      },
...
      "StaffSectionListResponse": {
        "properties": {
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/StaffSectionSummary"
            },
            "type": "array",
            "title": "Sections"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "sections"
        ],
        "title": "StaffSectionListResponse"
      },
      "StaffSectionSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "crn": {
            "type": "string",
            "title": "Crn"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "crn",
          "is_active"
        ],
        "title": "StaffSectionSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}